+++ test/unittests/cashflow_tests.py
@@ -3,7 +3,7 @@
 # dcf
 # ---
 # A Python library for generating discounted cashflows.
-# 
+#
 # Author:   sonntagsgesicht, based on a fork of Deutsche Postbank [pbrisk]
 # Version:  0.4, copyright Saturday, 10 October 2020
 # Website:  https://github.com/sonntagsgesicht/dcf
@@ -26,13 +26,20 @@ class CashflowListUnitTests(TestCase):
         self.amount = 111
 
     def test_str(self):
-        cf = self.cls(self.schedule)
+        cf = self.cls(self.schedule[:3])
 
         self.assertIn(self.cls.__name__, str(cf))
         self.assertIn(repr(cf.domain[0]), str(cf))
         self.assertIn(repr(cf.domain[-1]), str(cf))
+
+        self.assertIn(self.cls.__name__, repr(cf))
         self.assertIn(str(cf.domain), repr(cf))
-        self.assertIn(str(cf[cf.domain]), repr(cf))
+
+        fx = FixedCashFlowList(cf)
+
+        self.assertEqual(cf.domain, fx.domain)
+        self.assertIn(str(cf.domain), repr(fx))
+        self.assertIn(str(cf[cf.domain]), repr(fx))
 
         cf = self.cls(self.schedule, self.amount)